update styles, move into dedicated modules
