+++ app/[blog]/[post]/components/Reactions.module.css
@@ -1,0 +1,5 @@
+.buttons {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
